+++ CHANGELOG.rst
@@ -1,8 +1,10 @@
 Changelog
 =========
 
-Current (2025-01-24)
+Current (2025-01-29)
 --------------------
+
+* Implement a write_all_efficient function for ETROC2
 
 0.1.9 (2024-01-24)
 ------------------